// src/app/model/predicates/predicateTypes.ts
export enum PredicateType {
    true = 'true',
    id = 'id',
    equals = 'equal',
    not = 'not',
    date = 'date',
    greaterThen = 'gt',
    lessThen = 'lt',
    and = 'and',
    or = 'or',
    intAttr = 'int',
    strAttr = 'string',
    floatAttr = 'float',
    boolAttr = 'boolean',
    spread = 'spread'
}